requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
altair==4.2.0
appnope==0.1.3
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
base58==2.1.1
bleach==4.1.0
blinker==1.4
boto3==1.20.14
botocore==1.23.14
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
characteristic==14.3.0
charset-normalizer==2.0.8
click==7.1.2
cloudpickle==2.0.0
coloredlogs==15.0.1
ConfigArgParse==1.5.3
constantly==15.1.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
executing==0.8.3
fakeredis==1.6.1
fbmessenger==4.3.1
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.2
frozenlist==1.2.0
future==0.18.2
gast==0.4.0
gevent==21.8.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.18.2
greenlet==1.1.2
grpcio==1.42.0
h5py==3.6.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.8.2
incremental==21.3.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonpickle==2.0.0
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
klein==21.8.0
libclang==12.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
multitasking==0.0.10
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nbsphinx==0.8.7
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.11
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.5
pandoc==1.1.0
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.8.0
plumbum==1.7.1
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.10.0
pykwalify==1.6.0
Pympler==1.0.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
rasa==0.1.1
rasa-core==0.8.6
rasa-nlu==0.11.5
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.3
semver==2.13.0
Send2Trash==1.8.0
setuptools-scm==6.3.2
simplejson==3.17.6
six==1.16.0
slackclient==2.9.3
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==4.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
streamlit==1.9.2
streamlit-aggrid==0.2.3.post2
streamlit-multipage==0.0.18
streamlit-pages==0.1.0
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.15.0
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
Tubes==0.2.0
Twisted==21.7.0
typing==3.7.4.3
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.6.0
wrapt==1.13.3
yarl==1.7.2
yfinance==0.1.70
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
